Introduced basic multiprocessing * using python multiprocessing library * using a process pool with cpu_count processes * an example iterative implementation using itertools is also given
An ugly workaround is applied for unpickeable objects in module Physics: parts of objects are sent to the worker processes and then the object is reconstructed inside the worker process.
The code lacks many little but obvious performance improvements (e.g. using the same process pool for the entire simulation, remove unnecessary world creations in the main process etc.)

Anyway, it works.